requirements.txt
tgcrypto
pyrofork
uvloop
python-dotenv
sqlalchemy==1.3.23
pymongo==4.6.1
